add file: help_functions and modified file constants
